Remove no longer supported external_data option.
The introduction of the use of instances rather than classes for mappings makes passing external data into the importer unnecessary.  In cases where a mapping needs more data from the system running the imports, #get_mapper should be defined as an instance method for the mapping class and external data should be passed in when the instance is initialized.

```ruby
class FooMapper
  def initialize(external_data)
    @external_data = external_data
  end

  def get_mapper(_strategy_class)
    Topographer::Mapper.build_mapper(SomeClass) do |mapper|
      mapper.required_mapping([:some_column], :some_output_field) do |input_data|
        @input_data[:some_column] + @external_data[:some_modifier]
      end
    end
  end
end
```